requirements.txt
aiohttp==3.11.13
fastapi==0.115.12
Pillow==11.2.0
pydantic==2.11.1
uvicorn==0.34.0
websockets==15.0.1
